Rename ConnectMetaMask and add a proper connect button
