+++ src/components/MetaMaskAddChain/styles.module.css
@@ -1,0 +1,14 @@
+.button {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 0.6rem;
+  margin-bottom: 1em;
+  padding: 0.6rem;
+  border: none;
+  background: rgb(3, 125, 214);
+  border-radius: var(--ifm-global-radius);
+  cursor: pointer;
+  font-family: inherit;
+  font-size: inherit;
+}
