+++ src/theme/MDXComponents.ts
@@ -1,6 +1,7 @@
 // Import the original mapper
 import MDXComponents from '@theme-original/MDXComponents';
 import ConnectMetaMask from '@site/src/components/ConnectMetaMask';
+import MetaMaskAddChain from '@site/src/components/MetaMaskAddChain';
 import DiscordButton from '@site/src/components/DiscordButton';
 import RegisterReputationMining from '@site/src/components/RegisterReputationMining';
 
@@ -8,6 +9,7 @@ const components = {
   // Re-use the default mapping
   ...MDXComponents,
   'connect-metamask': ConnectMetaMask,
+  'metamask-add-chain': MetaMaskAddChain,
   'join-discord': DiscordButton,
   'register-reputation-mining': RegisterReputationMining,
 };
